Checking before shell call to remove temporary files in end_ensemble_manager
git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@1135 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 81d82c35cb9ff22be08bb6ab361cca17af344524

+++ ensemble_manager/ensemble_manager_mod.f90
@@ -617,10 +617,12 @@ if(mean_allocated) deallocate(ens_mean)
 if(spread_allocated) deallocate(ens_spread)
 
 ! LETS GET RID OF ALL THE TEMP FILES, TOO!
-command_string = 'rm -f ' // trim(ens_file_name) // '.*'
-call system(command_string)
-command_string = 'rm -f ' // trim(reg_file_name) // '.*'
-call system(command_string)
+if(.not. in_core) then
+   command_string = 'rm -f ' // trim(ens_file_name) // '.*'
+   call system(command_string)
+   command_string = 'rm -f ' // trim(reg_file_name) // '.*'
+   call system(command_string)
+endif
 
 end subroutine end_ensemble_manager
